# backend-laravel-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Receptarium  API",
        "contact": {
            "name": "David Pérez Romero",
            "email": "contacto@david-pr.com"
        },
        "license": {
            "name": "Licencia",
            "url": "http://www.example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth Controller"
                ],
                "summary": "Register a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test register"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "register@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth Controller"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "accessToken": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth Controller"
                ],
                "summary": "Logout user and revoke all tokens",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Successful logout"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pasos": {
            "post": {
                "tags": [
                    "Pasos receta Controller"
                ],
                "summary": "Store a step",
                "description": "Store a Step",
                "operationId": "9b53110247a6ec5649c6dfc06c8d6d04",
                "requestBody": {
                    "description": "Step data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasosRecetaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Step stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pasos/{id}": {
            "get": {
                "tags": [
                    "Pasos receta Controller"
                ],
                "summary": "Show a step by ID",
                "description": "Show step",
                "operationId": "1e0ad8fd07e1fd641caa689c9d2f21f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of step",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Step found"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pasos receta Controller"
                ],
                "summary": "Update step",
                "description": "Update  a step",
                "operationId": "a945971b00f8b8eccf6af4b9211cc0d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the step",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "step data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasosRecetaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Step stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pasos receta Controller"
                ],
                "summary": "Delete a step by ID",
                "description": "Delete a step",
                "operationId": "979dd8be6eebde9740d877a9321f10ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the step",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "203": {
                        "description": "Deleted recipe"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PasosRecetaRequest": {
                "properties": {
                    "titulo": {
                        "type": "string",
                        "example": "Paso 1"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripcion del paso 1"
                    },
                    "recetas_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ruta_imagen": {
                        "type": "string",
                        "example": "Ruta de la imagen , puede ser null"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}